// src/lib/grpc/gen/prosper/v1/ledger.ts
export interface LedgerAccount {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: prosper.v1.LedgerAccountType type = 3
     */
    type: LedgerAccountType;
    /**
     * Set only for ASSET accounts that mirror a BankAccount.
     *
     * @generated from protobuf field: optional int32 bank_account_id = 4
     */
    bankAccountId?: number;
}
/**
 * @generated from protobuf message prosper.v1.EntryLine
 */
export interface EntryLine {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: int32 transaction_id = 2
     */
    transactionId: number;
    /**
     * @generated from protobuf field: int32 ledger_account_id = 3
     */
    ledgerAccountId: number;
    /**
     * Exactly one of currency_code / stock_id is set.
     *
     * @generated from protobuf field: optional string currency_code = 4
     */
    currencyCode?: string;
    /**
     * @generated from protobuf field: optional int32 stock_id = 5
     */
    stockId?: number;
    /**
     * @generated from protobuf field: int64 amount_nanos = 6
     */
    amountNanos: bigint;
}
/**
 * @generated from protobuf message prosper.v1.SplitContext
 */
export interface SplitContext {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: int32 transaction_id = 2
     */
    transactionId: number;
    /**
     * @generated from protobuf field: string companion_name = 3
     */
    companionName: string;
    /**
     * @generated from protobuf field: int64 companion_share_nanos = 4
     */
    companionShareNanos: bigint;
    /**
     * @generated from protobuf field: int64 companion_paid_nanos = 5
     */
    companionPaidNanos: bigint;
}
// ...
export interface Transaction {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * Per-user incrementing ID, stable across edits.
     *
     * @generated from protobuf field: int32 iid = 2
     */
    iid: number;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp timestamp = 3
     */
    timestamp?: Timestamp;
    /**
     * @generated from protobuf field: string note = 4
     */
    note: string;
    /**
     * @generated from protobuf field: prosper.v1.TransactionType type = 5
     */
    type: TransactionType;
    /**
     * @generated from protobuf field: optional string vendor = 6
     */
    vendor?: string;
    /**
     * @generated from protobuf field: optional string payer = 7
     */
    payer?: string;
    /**
     * @generated from protobuf field: optional int32 category_id = 8
     */
    categoryId?: number;
    /**
     * @generated from protobuf field: optional int32 trip_id = 9
     */
    tripId?: number;
    /**
     * @generated from protobuf field: optional int32 supersedes_id = 10
     */
    supersedesId?: number;
    /**
     * @generated from protobuf field: bool is_void = 11
     */
    isVoid: boolean;
    /**
     * @generated from protobuf field: repeated int32 tag_ids = 12
     */
    tagIds: number[];
    /**
     * @generated from protobuf field: repeated prosper.v1.EntryLine lines = 13
     */
    lines: EntryLine[];
    /**
     * @generated from protobuf field: repeated prosper.v1.SplitContext splits = 14
     */
    splits: SplitContext[];
}
// ...
export enum TransactionType {
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_EXPENSE = 1;
     */
    EXPENSE = 1,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_INCOME = 2;
     */
    INCOME = 2,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_TRANSFER = 3;
     */
    TRANSFER = 3,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_THIRD_PARTY_EXPENSE = 4;
     */
    THIRD_PARTY_EXPENSE = 4,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_OPENING_BALANCE = 5;
     */
    OPENING_BALANCE = 5
}
/**
 * @generated from protobuf enum prosper.v1.LedgerAccountType
 */
export enum LedgerAccountType {
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_ASSET = 1;
     */
    ASSET = 1,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_EXPENSE = 2;
     */
    EXPENSE = 2,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_INCOME = 3;
     */
    INCOME = 3,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_EQUITY = 4;
     */
    EQUITY = 4,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_CURRENCY_EXCHANGE = 5;
     */
    CURRENCY_EXCHANGE = 5,
    /**
     * @generated from protobuf enum value: LEDGER_ACCOUNT_TYPE_RECEIVABLE = 6;
     */
    RECEIVABLE = 6
}
// ...
class LedgerAccount$Type extends MessageType<LedgerAccount> {
    constructor() {
        super("prosper.v1.LedgerAccount", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "type", kind: "enum", T: () => ["prosper.v1.LedgerAccountType", LedgerAccountType, "LEDGER_ACCOUNT_TYPE_"] },
            { no: 4, name: "bank_account_id", kind: "scalar", opt: true, T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<LedgerAccount>): LedgerAccount {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0;
        message.name = "";
        message.type = 0;
        if (value !== undefined)
            reflectionMergePartial<LedgerAccount>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LedgerAccount): LedgerAccount {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 id */ 1:
                    message.id = reader.int32();
                    break;
                case /* string name */ 2:
                    message.name = reader.string();
                    break;
                case /* prosper.v1.LedgerAccountType type */ 3:
                    message.type = reader.int32();
                    break;
                case /* optional int32 bank_account_id */ 4:
                    message.bankAccountId = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LedgerAccount, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 id = 1; */
        if (message.id !== 0)
            writer.tag(1, WireType.Varint).int32(message.id);
        /* string name = 2; */
        if (message.name !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.name);
        /* prosper.v1.LedgerAccountType type = 3; */
        if (message.type !== 0)
            writer.tag(3, WireType.Varint).int32(message.type);
        /* optional int32 bank_account_id = 4; */
        if (message.bankAccountId !== undefined)
            writer.tag(4, WireType.Varint).int32(message.bankAccountId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message prosper.v1.LedgerAccount
 */
export const LedgerAccount = new LedgerAccount$Type();
// ...
export const EntryLine = new EntryLine$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SplitContext$Type extends MessageType<SplitContext> {
    constructor() {
        super("prosper.v1.SplitContext", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "transaction_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "companion_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "companion_share_nanos", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 5, name: "companion_paid_nanos", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<SplitContext>): SplitContext {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0;
        message.transactionId = 0;
        message.companionName = "";
        message.companionShareNanos = 0n;
        message.companionPaidNanos = 0n;
        if (value !== undefined)
            reflectionMergePartial<SplitContext>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SplitContext): SplitContext {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 id */ 1:
                    message.id = reader.int32();
                    break;
                case /* int32 transaction_id */ 2:
                    message.transactionId = reader.int32();
                    break;
                case /* string companion_name */ 3:
                    message.companionName = reader.string();
                    break;
                case /* int64 companion_share_nanos */ 4:
                    message.companionShareNanos = reader.int64().toBigInt();
                    break;
                case /* int64 companion_paid_nanos */ 5:
                    message.companionPaidNanos = reader.int64().toBigInt();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SplitContext, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 id = 1; */
        if (message.id !== 0)
            writer.tag(1, WireType.Varint).int32(message.id);
        /* int32 transaction_id = 2; */
        if (message.transactionId !== 0)
            writer.tag(2, WireType.Varint).int32(message.transactionId);
        /* string companion_name = 3; */
        if (message.companionName !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.companionName);
        /* int64 companion_share_nanos = 4; */
        if (message.companionShareNanos !== 0n)
            writer.tag(4, WireType.Varint).int64(message.companionShareNanos);
        /* int64 companion_paid_nanos = 5; */
        if (message.companionPaidNanos !== 0n)
            writer.tag(5, WireType.Varint).int64(message.companionPaidNanos);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message prosper.v1.SplitContext
 */
export const SplitContext = new SplitContext$Type();
// ...
export const Transaction = new Transaction$Type();